women_support_forum/requirements.txt
APScheduler==3.10.4
Flask==3.0.3
Flask_APScheduler==1.13.1
Flask_Cors==5.0.0
Flask_JWT_Extended==4.6.0
Flask_Login==0.6.3
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.34
Werkzeug==3.0.4
